requirements.txt
alembic==1.7.5
apparmor==3.1.1
appdirs==1.4.4
attrs==22.1.0
autocommand==2.2.1
Babel==2.10.3
bcrypt==4.0.1
bidict==0.22.0
blinker==1.5
Brlapi==0.8.4
Brotli==1.0.9
btrfsutil==6.0
cffi==1.15.1
chardet==5.0.0
chrome-gnome-shell==0.0.0
click==8.1.3
cloud-init==22.3.4
configobj==5.1.0.dev0
cryptography==38.0.1
Cython==0.29.32
decorator==5.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
extras==1.0.0
fixtures==4.0.1
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
glfw==2.5.5
gpg==1.18.0
greenlet==1.1.3
gssapi==1.8.0
idna==3.4
imgui @ git+https://github.com/pyimgui/pyimgui.git@84bc6bcb9248fc3eb858ee59f144ee067e13ec24
importlib-metadata==5.0.0
inflect==6.0.2
itsdangerous==2.1.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.16.0
ldap3==2.9
lensfun==0.3.3
LibAppArmor==3.1.1
libfdt==1.6.1
louis==3.23.0
Mako==1.2.3
MarkupSafe==2.1.1
more-itertools==9.0.0
netifaces==0.11.0
netsnmp-python==1.0a1
nspektr==0.4.0
numpy==1.23.4
nyx==2.1.0
oauthlib==3.2.0
ordered-set==4.1.0
packaging==21.3
paramiko==2.11.0
passlib==1.7.4
pbr==5.11.0
ply==3.11
psutil==5.9.2
psycopg2==2.9.3
pwquality==1.4.4
pyasn1==0.4.8
pybind11==2.10.0
pycairo==1.21.0
pycparser==2.21
pydantic==1.10.2
pygame==2.1.2
PyGObject==3.42.2
PyNaCl==1.4.0
PyOpenGL==3.1.6
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.19.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.1
pyxdg==0.28
PyYAML==6.0
Reflector==2021.11.20.2.41.3
requests==2.28.1
simplejson==3.17.6
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.2
sshtunnel==0.4.0
stem==1.8.0
TBB==0.2
team==1.0
testtools==2.5.0
tomli==2.0.1
torbrowser-launcher==0.3.5
trove-classifiers==2022.10.19
typing_extensions==4.4.0
urllib3==1.26.12
validate==5.1.0.dev0
validate-pyproject==0.10.1
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
